requirements.txt
-i https://pypi.org/simple
amqp==2.4.2
beautifulsoup4==4.7.1
billiard==3.6.0.0
bs4==0.0.1
celery[redis]==4.3.0
certifi==2019.3.9
chardet==3.0.4
django-celery-beat==1.4.0
django-cors-headers==2.5.2
django-timezone-field==3.0
django==2.2
djangorestframework==3.9.2
idna==2.8
kombu==4.5.0
lxml==4.3.3
python-crontab==2.3.6
python-dateutil==2.8.0
pytz==2018.9
redis==3.2.1
requests==2.21.0
six==1.12.0
soupsieve==1.9
sqlparse==0.3.0
urllib3==1.24.1
vine==1.3.0
